feat: change page path about tos
비회원 접근 가능을 위한 tos 관련 페이지 경로 수정
