+++ src/app/(non-member)/signup/layout.tsx
@@ -1,0 +1,12 @@
+import { ReactNode } from 'react';
+
+import Header from '@/components/layout/Header';
+
+export default function Layout({ children }: { children: ReactNode }) {
+  return (
+    <>
+      <Header />
+      { children }
+    </>
+  );
+}
